Refactor JWT methods to services.js
