+++ user-service/services.js
@@ -1,4 +1,9 @@
 import bcrypt from 'bcrypt';
+import jwt from 'jsonwebtoken';
+import dotenv from 'dotenv';
+
+// get config vars
+dotenv.config();
 
 // hash and salt passwords
 export async function hashSaltPassword(password) {
@@ -14,4 +19,36 @@ export async function hashSaltPassword(password) {
  */
 export async function verifyPassword(password, hashed) {
     return bcrypt.compare(password, hashed);
+}
+
+export async function generateAccessToken(user) {
+    return jwt.sign(
+        { username: user.username },
+        process.env.ACCESS_TOKEN_SECRET,
+        { expiresIn: '30s' },
+    ); // expire in 30s
+}
+
+export async function generateRefreshAccessToken(user) {
+    return jwt.sign(
+        { username: user.username },
+        process.env.REFRESH_TOKEN_SECRET,
+        { expiresIn: '1800s' },
+    ); // expire in 30 mins
+}
+
+export async function verifyAccessToken(token) {
+    return jwt.verify(token, process.env.ACCESS_TOKEN_SECRET, (err, user) => {
+        console.log(err);
+        if (err) return undefined;
+        return user;
+    });
+}
+
+export async function verifyRefreshToken(refreshToken) {
+    return jwt.verify(refreshToken, process.env.REFRESH_TOKEN_SECRET, async (err, user) => {
+        if (err) return undefined;
+        const token = await generateAccessToken({ username: user.username });
+        return token;
+    });
 }
